Move value reference identities into core
