+++ crates/trust-runtime-core/src/value/mod.rs
@@ -3,5 +3,7 @@
 #![allow(missing_docs)]
 
 pub mod datetime;
+mod reference;
 
 pub use datetime::*;
+pub use reference::*;
