+++ crates/trust-runtime/src/value/reference.rs
@@ -1,81 +1,9 @@
-use smol_str::SmolStr;
-
-use crate::error::RuntimeError;
-use crate::memory::MemoryLocation;
-
 use super::{normalize_assignment_for_target, read_string_element, write_string_element, Value};
 
-/// Reference path segment within composite values.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum RefSegment {
-    Index(RefIndices),
-    Field(SmolStr),
-}
-
-pub type RefIndices = Vec<i64>;
-pub type RefPath = Vec<RefSegment>;
-
-/// Reference to a value in memory.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ValueRef {
-    pub location: MemoryLocation,
-    pub offset: usize,
-    pub path: RefPath,
-}
-
-pub(crate) fn ref_indices_from_iter<I>(indices: I) -> RefIndices
-where
-    I: IntoIterator<Item = i64>,
-{
-    indices.into_iter().collect()
-}
-
-#[inline]
-pub(crate) fn single_ref_index(index: i64) -> RefSegment {
-    RefSegment::Index(vec![index])
-}
-
-#[inline]
-pub(crate) fn array_offset_i64(dimensions: &[(i64, i64)], indices: &[i64]) -> Option<usize> {
-    if dimensions.len() != indices.len() {
-        return None;
-    }
-    let mut offset: i128 = 0;
-    let mut stride: i128 = 1;
-    for ((lower, upper), index) in dimensions.iter().zip(indices).rev() {
-        if index < lower || index > upper {
-            return None;
-        }
-        let lower = i128::from(*lower);
-        let upper = i128::from(*upper);
-        let index = i128::from(*index);
-        let len = upper.checked_sub(lower)?.checked_add(1)?;
-        let relative = index.checked_sub(lower)?;
-        offset = offset.checked_add(relative.checked_mul(stride)?)?;
-        stride = stride.checked_mul(len)?;
-    }
-    usize::try_from(offset).ok()
-}
-
-#[inline]
-pub(crate) fn checked_array_offset_i64(
-    dimensions: &[(i64, i64)],
-    indices: &[i64],
-) -> Result<usize, RuntimeError> {
-    if dimensions.len() != indices.len() {
-        return Err(RuntimeError::TypeMismatch);
-    }
-    for ((lower, upper), index) in dimensions.iter().zip(indices) {
-        if index < lower || index > upper {
-            return Err(RuntimeError::IndexOutOfBounds {
-                index: *index,
-                lower: *lower,
-                upper: *upper,
-            });
-        }
-    }
-    array_offset_i64(dimensions, indices).ok_or(RuntimeError::TypeMismatch)
-}
+pub use trust_runtime_core::value::{
+    array_offset_i64, checked_array_offset_i64, parse_partial_access, ref_indices_from_iter,
+    single_ref_index, PartialAccess, PartialAccessError, RefIndices, RefPath, RefSegment, ValueRef,
+};
 
 #[inline]
 pub(crate) fn read_value_path_borrowed<'a>(
@@ -205,51 +133,6 @@ fn single_string_index(indices: &[i64]) -> Option<i64> {
 fn write_string_path(text: &str, indices: &[i64], value: Value, wide: bool) -> Option<String> {
     let index = single_string_index(indices)?;
     write_string_element(text, index, value, wide).ok()
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PartialAccess {
-    Bit(u8),
-    Byte(u8),
-    Word(u8),
-    DWord(u8),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PartialAccessError {
-    IndexOutOfBounds { index: i64, lower: i64, upper: i64 },
-    TypeMismatch,
-}
-
-pub fn parse_partial_access(text: &str) -> Option<PartialAccess> {
-    let text = text.trim();
-    if text.is_empty() {
-        return None;
-    }
-    if let Some(stripped) = text.strip_prefix('%') {
-        let mut chars = stripped.chars();
-        let prefix = chars.next()?;
-        let digits: String = chars.collect();
-        let index = parse_access_index(&digits)?;
-        return match prefix.to_ascii_uppercase() {
-            'X' => Some(PartialAccess::Bit(index)),
-            'B' => Some(PartialAccess::Byte(index)),
-            'W' => Some(PartialAccess::Word(index)),
-            'D' => Some(PartialAccess::DWord(index)),
-            _ => None,
-        };
-    }
-    if text.chars().all(|c| c.is_ascii_digit() || c == '_') {
-        let index = parse_access_index(text)?;
-        return Some(PartialAccess::Bit(index));
-    }
-    None
-}
-
-fn parse_access_index(text: &str) -> Option<u8> {
-    let cleaned: String = text.chars().filter(|c| *c != '_').collect();
-    let value: u64 = cleaned.parse().ok()?;
-    u8::try_from(value).ok()
 }
 
 #[cfg(test)]
