+++ docs/internal/testing/checklists/architecture-improvements.md
@@ -43,6 +43,7 @@ Issue #51 init benchmark baseline evidence (2026-04-28):
 - [x] `ARCH-RTCORE-07` Move portable memory identity types (`MemoryLocation`, `IoArea`, `FrameId`, `InstanceId`) into `trust-runtime-core` while preserving `trust-runtime::memory` re-exports; validate with runtime-core tests, focused runtime memory tests, no-default core check, clippy for touched crates, and the full-map doctor dependency fence.
 - [x] `ARCH-RTCORE-08` Move portable date/time calculation helpers into `trust-runtime-core`, retain a private `trust-runtime` compatibility module for internal imports, and update the full-map top-level module policy so the removed physical `datetime` source module is not treated as stale architecture evidence.
 - [x] `ARCH-RTCORE-09` Move `RuntimeError` into `trust-runtime-core`, preserve `trust-runtime::error` as the public compatibility path, keep the core no-default check passing, and update the full-map top-level module policy so the removed physical `error` source module is not treated as stale architecture evidence.
+- [x] `ARCH-RTCORE-10` Move portable value reference identities, array-offset helpers, and partial-access parsing into `trust-runtime-core`, preserve the `trust-runtime::value::reference` compatibility re-export, and leave `Value`-dependent path walkers in the host crate until the value model ownership boundary is split by accessor and mutation responsibilities.
 - [x] `ARCH-FULL-01` Add `full-architecture-refactor-program-checklist.md` as the umbrella board that states the runtime-core split is one branch of the full SOLID/KISS and zero-silent-bug program, not the whole program.
 - [x] `ARCH-FULL-02` Add `architecture-doctor-full-map-execution-checklist.md` so `architecture-doctor --full-map` becomes a concrete prerequisite for enforcing workspace, runtime-core, CLI, host-surface, KISS, API, and diagram rules.
 - [x] `ARCH-FULL-03` Add `hir-mutation-hardening-execution-checklist.md` for the F2 silent-bug gap: `symbol_import`, `type_check::const_eval`, and aggregate initializer validation mutation survivors.
