+++ docs/internal/testing/checklists/runtime-core-host-split-execution-checklist.md
@@ -206,6 +206,10 @@ Move lowest-risk portable pieces first.
 - [x] `RTSPLIT-P4-ERROR-001` Move the shared runtime error type into `trust-runtime-core`. Evidence: `crates/trust-runtime-core/src/error.rs` now owns `RuntimeError`; `trust-runtime::error` re-exports the core error API so existing external and internal paths keep compiling.
 - [x] `RTSPLIT-P4-ERROR-002` Verify error compatibility after the move. Evidence: `cargo test -p trust-runtime --test errors_policy --test runtime_core_behavior_lock -- --nocapture` passes 7 tests; `cargo test -p trust-runtime-core -- --nocapture`, `cargo check -p trust-runtime --lib`, `cargo check -p trust-runtime-core --no-default-features`, and `cargo clippy -p trust-runtime-core -p trust-runtime --lib -- -D warnings` pass.
 - [x] `RTSPLIT-P4-ERROR-003` Keep the dependency/import fence green after the error move. Evidence: `cargo run -p xtask -- architecture-doctor --full-map` passes after removing the stale `error` physical top-level module decision; `FULLMAP-CHECK-05` still reports 18 forbidden dependencies and 17 forbidden import modules with no findings.
+- [x] `RTSPLIT-P4-REF-001` Move portable value reference identities and path helpers into `trust-runtime-core`. Evidence: `crates/trust-runtime-core/src/value/reference.rs` now owns `RefSegment`, `RefIndices`, `RefPath`, `ValueRef`, `ref_indices_from_iter`, `single_ref_index`, `array_offset_i64`, and `checked_array_offset_i64`; `trust-runtime::value::reference` re-exports the moved API and keeps only the `Value`-dependent path walkers host-side.
+- [x] `RTSPLIT-P4-REF-002` Move portable partial-access parsing into `trust-runtime-core` without moving the whole value model. Evidence: `PartialAccess`, `PartialAccessError`, and `parse_partial_access` now live in the core reference module, while string/array/struct mutation still stays in `trust-runtime` until `Value` field accessors and ownership boundaries are split cleanly.
+- [x] `RTSPLIT-P4-REF-003` Verify the reference slice with focused tests and no-default compatibility. Evidence: `cargo test -p trust-runtime-core -- --nocapture` passes 12 tests, `cargo test -p trust-runtime value::reference --lib -- --nocapture` passes 3 focused runtime reference tests, `cargo check -p trust-runtime --lib` passes, `cargo check -p trust-runtime-core --no-default-features` passes, and `cargo clippy -p trust-runtime-core -p trust-runtime --lib -- -D warnings` passes.
+- [x] `RTSPLIT-P4-REF-004` Keep the dependency/import fence green after the reference move. Evidence: `cargo run -p xtask -- architecture-doctor --full-map` passes; `FULLMAP-CHECK-05` still reports 18 forbidden dependencies and 17 forbidden import modules with no findings.
 
 ### Phase 4 Exit Gate
